Mod directory support
The mod will now search for json files under the directory
`Mods/PPDefModifier`, if it exists. Every `.json` file in this directory
or a subdirectory of it will be loaded, allowing for better management
of mod configuration files.

The original `PPDefModifier.json` file directly in the `Mods` directory
is also still supported, if present.